Using and saving the randomly sampled noise from cfg.noise
